+++ src/gen-charts.ts
@@ -570,16 +570,20 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 
 		// OPTION: Title
 		if (rel.opts.showTitle) {
-			strXml += genXmlTitle({
-				title: rel.opts.title || 'Chart Title',
-				color: rel.opts.titleColor,
-				fontFace: rel.opts.titleFontFace,
-				fontSize: rel.opts.titleFontSize || DEF_FONT_TITLE_SIZE,
-				titleAlign: rel.opts.titleAlign,
-				titleBold: rel.opts.titleBold,
-				titlePos: rel.opts.titlePos,
-				titleRotate: rel.opts.titleRotate,
-			})
+			strXml += genXmlTitle(
+				{
+					title: rel.opts.title || 'Chart Title',
+					color: rel.opts.titleColor,
+					fontFace: rel.opts.titleFontFace,
+					fontSize: rel.opts.titleFontSize || DEF_FONT_TITLE_SIZE,
+					titleAlign: rel.opts.titleAlign,
+					titleBold: rel.opts.titleBold,
+					titlePos: rel.opts.titlePos,
+					titleRotate: rel.opts.titleRotate,
+				},
+				rel.opts.x as number,
+				rel.opts.y as number
+			)
 			strXml += '<c:autoTitleDeleted val="0"/>'
 		} else {
 			// NOTE: Add autoTitleDeleted tag in else to prevent default creation of chart title even when showTitle is set to false
@@ -1977,15 +1981,25 @@ function makeSerAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
  * @param {IChartPropsTitle} opts - options
  * @return {string} XML `<c:title>`
  */
-function genXmlTitle(opts: IChartPropsTitle): string {
+function genXmlTitle(opts: IChartPropsTitle, chartX?: number, chartY?: number): string {
 	let align = opts.titleAlign === 'left' || opts.titleAlign === 'right' ? `<a:pPr algn="${opts.titleAlign.substring(0, 1)}">` : `<a:pPr>`
 	let rotate = opts.titleRotate ? `<a:bodyPr rot="${convertRotationDegrees(opts.titleRotate)}"/>` : `<a:bodyPr/>` // don't specify rotation to get default (ex. vertical for cat axis)
 	let sizeAttr = opts.fontSize ? 'sz="' + Math.round(opts.fontSize * 100) + '"' : '' // only set the font size if specified.  Powerpoint will handle the default size
 	let titleBold = opts.titleBold === true ? 1 : 0
-	let layout =
-		opts.titlePos && opts.titlePos.x && opts.titlePos.y
-			? `<c:layout><c:manualLayout><c:xMode val="edge"/><c:yMode val="edge"/><c:x val="${opts.titlePos.x}"/><c:y val="${opts.titlePos.y}"/></c:manualLayout></c:layout>`
-			: `<c:layout/>`
+
+	let layout = '<c:layout/>'
+	if (opts.titlePos && typeof opts.titlePos.x === 'number' && typeof opts.titlePos.y === 'number') {
+		// NOTE: manualLayout x/y vals are *relative to entire slide*
+		const totalX = opts.titlePos.x + chartX
+		const totalY = opts.titlePos.y + chartY
+		let valX = totalX === 0 ? 0 : (totalX * (totalX / 5)) / 10
+		if (valX >= 1) valX = valX / 10
+		if (valX >= 0.1) valX = valX / 10
+		let valY = totalY === 0 ? 0 : (totalY * (totalY / 5)) / 10
+		if (valY >= 1) valY = valY / 10
+		if (valY >= 0.1) valY = valY / 10
+		layout = `<c:layout><c:manualLayout><c:xMode val="edge"/><c:yMode val="edge"/><c:x val="${valX}"/><c:y val="${valY}"/></c:manualLayout></c:layout>`
+	}
 
 	return `<c:title>
 	  <c:tx>